requirements.txt
# Database
psycopg2-binary>=2.9.5

# Data processing
numpy>=1.20.0
pandas>=1.3.0

# Machine Learning
scikit-learn>=1.0.0

# Visualization
plotly>=5.10.0

# Optional: Jupyter support for interactive visualizations
jupyter>=1.0.0
ipywidgets>=8.0.0

python_dotenv>=1.0.1
